Cache classloaders of kapt when running in Gradle workers
The loaders and hence some classes were repeatedly loaded and
jit-compiled everytime when KaptExecution were dispatched. Those
classes, like JavaCompiler, can be very large and therefore created a
significant overhead. In some projects, the overhead accounted for more
than 40% of total CPU time of annotation processing.

This change tries to cache the classloaders in a WeakHashMap so that
whenever there are enough memory, they won't be reloaded and re-jitted.