Add support for Closure library modules.
This makes it much easier to integrate with existing codebases using
closure libraries and the closure compiler.
See https://github.com/google/closure-library/wiki/goog.module:-an-ES6-module-like-alternative-to-goog.provide
for more info.

Note that this doesn't currently work well with DCE, but the intent is to
compile the output with closure anyway.