Generate synthetic functions for local functions with default values, and invoke them with the correct function descriptor.
This is accomplished by having DefaultArgumentStubGenerator transform StatementContainers (local function declarations are statements) in addition to DeclarationContainers.

Without the changes to AsmUtil, the generated functions are not invoked with the correct descriptor; the enclosing class is included in the descriptor as the dispatch receiver parameter.

This change still does not fix local functions with default values from closures (i.e., KT-5589).