Build the compose ide plugin
Moves the sources from the old structure to the new
AndroidX structure. Builds a working version of the
ide plugin as well as builds a version of the compiler
plugin that is intended to be used in the ide.

Test: ./gradle test
Change-Id: I68962d3dd228d4bf6e10abc8e5aec5228042b038 ( https://android-review.googlesource.com/q/I68962d3dd228d4bf6e10abc8e5aec5228042b038 )

Moved from: https://github.com/androidx/androidx/commit/63d995f17847047e641f4af8f4fd985892d63444