Only run the compose plugin tests once
There is no need to run both release and debug tests as there
is only a one version of the plugin jar created.

The debug tests are chosen as it is simpler to debug the tests
themselves in the ide.

Test: ./gradlew test
Change-Id: Ifbb6baedfc2ef730096d36c7ad5e0a34cbce0c52 ( https://android-review.googlesource.com/q/Ifbb6baedfc2ef730096d36c7ad5e0a34cbce0c52 )

Moved from: https://github.com/androidx/androidx/commit/0fd24822c7290a525bdf54cc639ef72e827e6729