StatisticsUpdate: Use a weak-ref to DocumentListener disposable
The Document could be associated with an Editor of some text field,
and code using it might expect its life-time to be limited within the
Project scope, thus not cleaning up its listeners attached to the
document. But when StatisticsUpdate (an application-scoped component)
kicks in, it adds its own listener to the document and registers a
Disposable to remove the listener.

However this in turn adds a hard-ref from the Disposer to the document
and all its other listeners, some of which expect the Document instance
to be gone by the time the project is closed.

While this approach is inherently broken, this change makes the Project
leak tests pass regardless any improper usage of Document listeners.
In real life the StatisticsUpdate performs the necessary cleanup after
some time anyway.

intellij-monorepo: 1c4d9a6d147a860b3a60b605a1f0ecf5b225ff94