Fixes nullable @Composable type checking
`null` cannot be @Composable, so we need an escape hatch for
trying to assign null to a nullable @Composable type, such as:

val foo: @Composable() (() -> Unit)? = null

It is safe for us to just check if the type is nullable, and if the
expression type is Nothing? - if both are true then there's no need to
check for the @Composable annotation.

Bug: b/137311217 ( https://issuetracker.google.com/issues/137311217 )
Test: ComposableCheckerTests, demos
Change-Id: I06d68087feced91c41547d945c1eba91fd79726b ( https://android-review.git.corp.google.com/q/I06d68087feced91c41547d945c1eba91fd79726b )