JVM_IR: cast field receiver to owner type
This is necessary because the descriptors for fake overrides
incorrectly (?) have dispatch receiver type equal to the substituted
type of the real implementation's owner, not the actual owner. Since
psi2ir uses descriptors to generate IR fake overrides and insert
implicit casts, the resulting IR contains incorrect casts, e.g. in

    open class A { fun f() = 1 }
    class B : A { fun g(x: Any) = if (x is B) x.f() else null }

the call to `f` points to the fake override `B::f`, but `x` is casted to
`A` instead of `B`. While for functions, like in the example above, this
problem is masked in ExpressionCodegen by casting the receiver (the
bytecode contains two casts in a row), this was not done for fields.