Gradle: Escape TC service messages in debug logs
The MPP plugin uses TC service messages to interact with JS and
native test runs an writes received messages to a debug log. But
if a build is executed by TeamCity, these log messages may be
treated by TC as actual service messages.

This patch transforms the logged messages to avoid this issue.