RUBY-25512 add text-based resolve to "super" method calls in ruby modules
There is a design pattern where a module can be used to "extend" behaviour to methods by overriding it, doing something (e.g., logging) and calling super at some point. Unfortunately we cannot tell what "super" refers to as the module doesn't actually override anything yet. This allows us to offer all possible definitions of it as a safety net.

This required a tiny update to GotoDeclarationAction to ensure we can "choose ambiguous targets" when invoking resolve on keywords, as in the ruby language "super" is a keyword but it's also a method call (that can resolve somewhere).

intellij-monorepo: 8d9d6ddfafa246f0bdcc918c4b063a0731ee0a6c