CPP-16872: Use base of deferred icon for width without access icon
To properly cut the access specifier icon part from the deferred
icon we need to know if the deferred icon content is of RowIcon type.

The way of doing that with the least amount of side effects is to get
the base icon when we calculate the icon width.

intellij-monorepo: 1005fb175bb62d8fe3a7599016d79fc9248e0a00