gradle task enhancer: properly escape jvmArgs on Windows
Gradle use `java.lang.ProcessBuilder` to start external processes and pass `jvmArgs` to the builder "as is". However it is known, that escaping of the command line arguments used in `ProcessBuilder` is broken.
Our escaping implementation is used to workaround those problems.

this is also a fix for IDEA-222512

intellij-monorepo: 1325620a1db204b280c8bc3b76c62354f54029bc