IDEA-229109, RUBY-25512 ensure "Go to Declaration" for keywords works when there are ambiguous targets
The current behaviour allows go to declaration on keywords if they resolve to a single target, however for multiple targets it does nothing. Moving this short-circuit lower down allows `chooseAmbiguousTarget` to be called for keywords but still prevents the error message showing if no target is found.

This was fixed in 66e40617fef4401be106137c4fd219f31c601ecb, but due to a bad merge it was accidentally reverted, now it's back again.

intellij-monorepo: d293609a36a5aea3f6799f733bf776bd715759ec