Use more robust way to find bundled binaries
Previously, we've expected 'PathManager.getPluginsPath()' to return path
to folder with IDE plugins installed. That's actually might not be the
case, as there are some environment variables (like 'plugin.path') which
might override installation folder of plugins without affecting
PathManager.

So far, the only known such case is running a debug IDE in Kotlin build.

Still, we've decided not to rely on invariants about locations of IDEA
plugins, and, instead, realy on the layout of our plugin: now we look
for mobile-jar plugin, and look for bundled binaries relative from that.
This should be more robust because we have control over this layout
during plugin build.