Make compose-runtime expect/actuals internal
All the expect/actuals used in Compose-runtime are implementation
details so they shouldn't be public.

I changed to direct import "android." for the few uses of them in ui/.
For the trace {} convenience wrapper, Leland suggested it could be
public but in an utility module, so I put one in androidx.ui.util.

Fixes: 149526684 ( https://issuetracker.google.com/issues/149526684 )
Test: cd ui/; ./gradlew dokkaKotlinTipOfTreeDocs
Change-Id: I215d024fa10e6a6bab128ad0deaa2024437941fa ( https://android-review.git.corp.google.com/q/I215d024fa10e6a6bab128ad0deaa2024437941fa )