ShowUsagesAction: don't call #collectData with empty lists
"Searching..." node is returned only when both lists are empty. This happens when search is not started yet.
Next time #collectData is called from #rebuildTable, which in turn is called from PingEDT (if popup is still showing),
which is pinged:
- when some result arrives, meaning the usages list is not empty at that point;
- when the search is finished; popup is cancelled if usages list is empty before PingEDT invokes its action.

intellij-monorepo: d9c71307efb53dbb8810473a6cff3d102fe5403a