Composition refactoring
Composition is an interface now and it has no compose method as we return this object from Activity.setContent{} and the returned object shouldn't have ability to compose any composable as this means this composable will override all the ambients we provided.
UiComposition is not needed anymore and there is no public Composition implementations at all. Instead there is a function compose() in compose-runtime which is used everytime we need to compose anything: View, Compose UI composables, subcompositions, Vectors. This also means HotReload hocks will be working for any types of compositions, previously it wasn't working for Vectors.
Postponed composition for the cases when we wait for the view's state to be restored is implemented with just composition emptyContent() while we are not ready and then recomposing with the real content, without any api changes required on the Composition side.
Deprecated methods used to compose/dispose are removed to simplify the refactoring. They were released as deprecated during the last release, so I hope it is ok to remove them now.

Test: manually, tests are passing
Bug: 151618508 ( https://issuetracker.google.com/issues/151618508 )
Change-Id: Ic52b58110c6c0fcf10d449c78e2cb7ba92c92b07 ( https://android-review.git.corp.google.com/q/Ic52b58110c6c0fcf10d449c78e2cb7ba92c92b07 )